Implementing the feedbacks

+++ app/src/main/java/com/tlrm/mobile/whapp/mvvm/picklist/viewmodel/PickListViewModel.kt
@@ -3,16 +3,20 @@ package com.tlrm.mobile.whapp.mvvm.picklist.viewmodel
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import androidx.lifecycle.*
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.tlrm.mobile.whapp.mvvm.picklist.model.PickListItem
 import com.tlrm.mobile.whapp.mvvm.picklistdetails.view.PickListDetailsActivity
 import com.tlrm.mobile.whapp.services.PickListService
+import com.tlrm.mobile.whapp.services.SessionService
 import com.tlrm.mobile.whapp.util.LoadingState
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
 
 class PickListViewModel(private val context: Context,
+                        private val sessionService: SessionService,
                         private val pickListService: PickListService): ViewModel() {
 
     private val TAG = PickListViewModel::class.java.simpleName
@@ -58,6 +62,11 @@ class PickListViewModel(private val context: Context,
     fun deletePickList(pickListNumber: String) {
         viewModelScope.launch {
             _loadingState.value = LoadingState.loading("Deleting picklist")
+            val user = sessionService.getUser()
+            if (!user.roles.contains("Mobile Manager")) {
+                _loadingState.value = LoadingState.error("You don't have authorize to delete pick list")
+                return@launch;
+            }
             try {
                 pickListService.deletePickList(pickListNumber)
                 _loadingState.value = LoadingState.LOADED
